requirements.txt
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.3
django-formtools==2.4
djangorestframework==3.14.0
executing==1.2.0
gunicorn==20.1.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
